requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
openai==1.12.0
langchain==0.1.0
langchain-core==0.1.10
langchain-openai==0.0.2
langgraph==0.0.20
pydantic==2.5.0
tiktoken==0.5.2